Clean up references to answer distribution-like things
In the near future an answer distribution report will be displayed in the data download reports table.  In order to avoid confusion, we need to clean up several misleading labels and remove a redundant feature.

Note that the redundant feature is surrounded by a feature flag so that it can be disabled once the report goes live in the table.

Fixes: AN-694